Ticket #1548: SK modules: change key lifetime type to uint64_t
In OPC UA specification lifetime is managed as Double/Duration type which is compatible with uint64_t use in API to allow longer periods than with uint32_t

+++ samples/ClientServer/demo_server/sks_demo_server_methods.c
@@ -132,8 +132,8 @@ SOPC_StatusCode SOPC_Method_Func_PublishSubscribe_GetSecurityKeys(const SOPC_Cal
     uint32_t FirstTokenId = 0;
     SOPC_ByteString* Keys = NULL;
     uint32_t NbToken = 0;
-    uint32_t TimeToNextKey = 0;
-    uint32_t KeyLifetime = 0;
+    uint64_t TimeToNextKey = 0;
+    uint64_t KeyLifetime = 0;
 
     SOPC_SKManager* skm = SOPC_SK_SecurityGroup_GetSkManager(SOPC_String_GetRawCString(&inputArgs[0].Value.String));
     SOPC_ReturnStatus status =
